+++ Packages/com.unity.inputsystem/InputSystem/Utilities/OneOrMore.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections;
 using System.Collections.Generic;
 
 namespace UnityEngine.InputSystem.Utilities
@@ -8,7 +9,7 @@ namespace UnityEngine.InputSystem.Utilities
     /// having to allocate GC heap garbage or having to alternatively split code paths.
     /// </summary>
     /// <typeparam name="TValue"></typeparam>
-    internal struct OneOrMore<TValue, TList>
+    internal struct OneOrMore<TValue, TList> : IReadOnlyList<TValue>
         where TList : IReadOnlyList<TValue>
     {
         private readonly bool m_IsSingle;
@@ -52,6 +53,42 @@ namespace UnityEngine.InputSystem.Utilities
         public static implicit operator OneOrMore<TValue, TList>(TList multiple)
         {
             return new OneOrMore<TValue, TList>(multiple);
+        }
+
+        public IEnumerator<TValue> GetEnumerator()
+        {
+            return new Enumerator { m_List = this };
+        }
+
+        IEnumerator IEnumerable.GetEnumerator()
+        {
+            return GetEnumerator();
+        }
+
+        private class Enumerator : IEnumerator<TValue>
+        {
+            internal int m_Index = -1;
+            internal OneOrMore<TValue, TList> m_List;
+
+            public bool MoveNext()
+            {
+                if (m_Index >= m_List.Count)
+                    return false;
+                ++m_Index;
+                return true;
+            }
+
+            public void Reset()
+            {
+                m_Index = -1;
+            }
+
+            public TValue Current => m_List[m_Index];
+            object IEnumerator.Current => Current;
+
+            public void Dispose()
+            {
+            }
         }
     }
 }